\name{RInside-package}
\alias{RInside-package}
\alias{RInside}
\docType{package}
\title{Embedding R in C++ applications}
\description{The \pkg{RInside} package makes it easier to embed R
  in your C++ applications. There is no code you would execute directly
  from the \code{R} environment. Rather, you write \code{C++} programs
  that embed \code{R} which is illustrated by some the included examples.
}
\author{Dirk Eddelbuettel and Romain Francois}
\keyword{programming}
\keyword{interface}
